refactor: merge common package to containers
