--- dagger/common/common.go
@@ -1,6 +1,0 @@
-package common
-
-import "errors"
-
-// ErrMissingRequiredArgument is returned when a required argument is missing.
-var ErrMissingRequiredArgument = errors.New("missing required argument")
