--- dagger/common/doc.go
@@ -1,2 +1,0 @@
-// Package common provides common functionality for the Dagger.
-package common
